% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_physics.R
\name{energy_bow}
\alias{energy_bow}
\title{Calculate energy of drawn bow}
\usage{
energy_bow(draw_weight_pounds, draw_length_inches)
}
\arguments{
\item{draw_weight_pounds}{draw weight in pounds}

\item{draw_length_inches}{draw length in inches}
}
\description{
The energy content of a bow with a drawn string is approximated with a linear
increase of the draw weight per drawn length.
}
\examples{
draw_weight <- 20 # pounds
draw_length <- 28 # inches
energy_bow(draw_weight, draw_length)
}
