% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_physics.R
\name{flight_arrow}
\alias{flight_arrow}
\title{Calculate flight of arrow}
\usage{
flight_arrow(initial_velocity, angle_degrees)
}
\arguments{
\item{angle_degrees}{}
}
\description{
The flight of the arrow is calculated with a given initial velocity of the arrow
at string release and the release angle.
}
