% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_physics.R
\name{init_v}
\alias{init_v}
\title{Calculate the initial velocity of an arrow at string release}
\usage{
init_v(E_pot, m)
}
\arguments{
\item{E_pot}{potential energy}

\item{m}{arrow weight in kg}
}
\description{
Calculate the initial velocity of an arrow at string release
}
