% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_physics.R
\name{plot_flight}
\alias{plot_flight}
\title{Plot the flight of an arrow}
\usage{
plot_flight(flight_data, max_x = NA, max_y = NA, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot the flight of an arrow
}
